Modify_StatHandler,Stat

+++ Assets/Scripts/ScriptableObject/StatScriptableObject.cs
@@ -12,7 +12,7 @@ public class StatScriptableObject : ScriptableObject
     public string characterName;
     public string description;
     public int maxSpeed;
-    public float weight;
+    public int weight;
     public int maxHP;
     public int maxMP;
 }
